+++ python_jobs/src/hbnmigration/config.py
@@ -2,6 +2,21 @@
 
 import os
 from pathlib import Path
+from typing import cast
+
+from .utility_functions.datatypes import ProjectStatus, ProjectStatuses
+
+
+def _get_project_status() -> ProjectStatus:
+    """Get and validate project status from environment."""
+    _project_status = os.environ.get("HBNMIGRATION_PROJECT_STATUS", "prod")
+    if _project_status not in ProjectStatuses:
+        msg = (
+            f"${{HBNMIGRATION_PROJECT_STATUS}} must be one of {ProjectStatuses} but is "
+            f"set to {_project_status}"
+        )
+        raise ValueError(msg)
+    return cast(ProjectStatus, _project_status)
 
 
 class Config:
@@ -13,3 +28,14 @@ class Config:
         else NotImplemented
     )
     """Path to root of hbnmigration project on server."""
+
+    PROJECT_STATUS: ProjectStatus = _get_project_status()
+    """
+    Project status.
+
+    dev = HBN - Intake and Curious (TEMP for Transition) PID 744
+    prod = HBN - Operations and Data Collection PID 625
+    """
+
+
+__all__ = ["Config"]