+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.7.0
+
+### Changed
+
+- Moved project status to envrironment variable `$HBNMIGRATION_PROJECT_STATUS` with default `"prod"`
+
 ## 1.6.1
 
 ### Added